ml_pipeline/requirements.txt
torch
torch-geometric
ogb
wandb
tqdm
dotenv
aiohttp
pandas
google-generativeai
faiss-cpu